requirements.txt
ipython==8.26.0
pillow==10.4.0
django-rest-marshmallow==4.0.2
marshmallow==3.21.3
django-crispy-forms==2.3
marshmallow==3.21.3